services/jwt_service/requirements.txt
flask==2.3.3
redis==4.6.0
python-dotenv==1.0.0
gunicorn==21.2.0